feat: add to bot rule

+++ src/Bot/QQBot.cs
@@ -24,7 +24,7 @@ public sealed class QqBot : IBot
         _bot = BotFactory.Create(new(), deviceInfo, _keyStore ?? new BotKeystore());
     }
 
-    public long Id => _bot.BotUin;
+    public uint Id => _bot.BotUin;
 
     public async Task SendMessageAsync(MessagePair messages, BotMessage source, bool noReply)
     {
